+++ lib/feature/status/widget/cli_widget.dart
